setting up react client development

+++ rankit/src/main/java/rankit/model/Player.java
@@ -41,7 +41,7 @@ public class Player {
 	public int getPoints() {
 		return points;
 	}
-	public void addPoints(int points) {
+	public void addPoints(double points) {
 		this.points += points;
 	}
 	public int getLostGames() {
